fix: remove memo and useCallback don't need
